backend/requirements.txt
# ============================================
# Core
# ============================================
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# ============================================
# Database
# ============================================
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
greenlet==3.1.1

# ============================================
# Authentication
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# ============================================
# File Parsing
# ============================================
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
python-magic==0.4.27  # MIME detection (use python-magic-bin on Windows)

# ============================================
# AI
# ============================================
openai==1.58.1
anthropic==0.42.0

# ============================================
# Task Queue
# ============================================
celery[redis]==5.4.0
redis==5.2.1

# ============================================
# Storage
# ============================================
supabase==2.11.0

# ============================================
# Utilities
# ============================================
httpx==0.28.1
pydantic[email]
python-dotenv==1.0.1

# ============================================
# Dev / Testing
# ============================================
pytest==8.3.4
pytest-asyncio==0.25.0
ruff==0.8.6
